requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.38.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
plotly==5.11.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
redis==4.3.4
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
spotipy==2.21.0
sqlparse==0.4.3
tenacity==8.1.0
threadpoolctl==3.1.0
tzdata==2022.6
urllib3==1.26.12
wrapt==1.14.1
